% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_config.R
\name{adapter_from_config}
\alias{adapter_from_config}
\title{Create Database Adapter from Config File}
\usage{
adapter_from_config(config_path = "./config.yml")
}
\arguments{
\item{config_path}{Path to config.yml file (default: "./config.yml")}
}
\value{
DatabaseAdapter subclass instance
}
\description{
Convenience function that loads configuration and creates the appropriate
database adapter in one step.
}
\examples{
\dontrun{
  adapter <- adapter_from_config("config.yml")
  conn <- adapter$connect()
  # ... use connection ...
  adapter$disconnect(conn)
}

}
\keyword{internal}
