% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\name{add_data_category}
\alias{add_data_category}
\title{Add Data Category}
\usage{
add_data_category(
  assessment_id,
  data_category,
  is_special_category = FALSE,
  data_subjects = NULL,
  retention_period = NULL,
  source_of_data = NULL
)
}
\arguments{
\item{assessment_id}{Assessment ID}

\item{data_category}{Category}

\item{is_special_category}{Whether special category}

\item{data_subjects}{Description of data subjects}

\item{retention_period}{Retention period}

\item{source_of_data}{Source}
}
\value{
List with success status
}
\description{
Add Data Category
}
\keyword{internal}
