% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{add_field_option}
\alias{add_field_option}
\title{Add Field Option}
\usage{
add_field_option(
  field_id,
  option_value,
  option_label,
  option_order = 1,
  is_default = FALSE
)
}
\arguments{
\item{field_id}{Field ID}

\item{option_value}{Option value}

\item{option_label}{Option label}

\item{option_order}{Order}

\item{is_default}{Whether default}
}
\value{
List with success status
}
\description{
Add Field Option
}
\keyword{internal}
