% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{add_impact_assessment}
\alias{add_impact_assessment}
\title{Add Impact Assessment}
\usage{
add_impact_assessment(
  request_id,
  impact_area,
  impact_level,
  assessed_by,
  impact_description = NULL,
  mitigation_strategy = NULL,
  affected_subjects = 0,
  affected_records = 0,
  requires_revalidation = FALSE
)
}
\arguments{
\item{request_id}{Request ID}

\item{impact_area}{Area of impact}

\item{impact_level}{Level of impact}

\item{assessed_by}{Assessor}

\item{impact_description}{Description}

\item{mitigation_strategy}{Mitigation strategy}

\item{affected_subjects}{Number of affected subjects}

\item{affected_records}{Number of affected records}

\item{requires_revalidation}{Whether revalidation required}
}
\value{
List with success status
}
\description{
Add Impact Assessment
}
\keyword{internal}
