% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{add_implementation_step}
\alias{add_implementation_step}
\title{Add Implementation Step}
\usage{
add_implementation_step(request_id, implementation_step, step_order = 1)
}
\arguments{
\item{request_id}{Request ID}

\item{implementation_step}{Step description}

\item{step_order}{Order}
}
\value{
List with success status
}
\description{
Add Implementation Step
}
\keyword{internal}
