% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{add_portability_dataset}
\alias{add_portability_dataset}
\title{Add Portability Dataset}
\usage{
add_portability_dataset(
  request_id,
  table_name,
  data_category,
  data_source,
  added_by,
  record_count = 0,
  is_provided_by_subject = TRUE,
  is_derived = FALSE
)
}
\arguments{
\item{request_id}{Request ID}

\item{table_name}{Table containing the data}

\item{data_category}{Category of data}

\item{data_source}{Source type from get_data_source_types()}

\item{added_by}{User adding the dataset}

\item{record_count}{Number of records}

\item{is_provided_by_subject}{Whether data was provided by subject}

\item{is_derived}{Whether data is derived/inferred}
}
\value{
List with success status and dataset details
}
\description{
Adds a dataset to be considered for portability export.
}
\examples{
\dontrun{
dataset <- add_portability_dataset(
  request_id = 1,
  table_name = "subjects",
  data_category = "CONTACT",
  data_source = "PROVIDED",
  added_by = "dpo"
)
}
}
\keyword{internal}
