% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\name{add_timeline_event}
\alias{add_timeline_event}
\title{Add Timeline Event}
\usage{
add_timeline_event(incident_id, event_type, event_description, recorded_by)
}
\arguments{
\item{incident_id}{Incident ID}

\item{event_type}{Event type}

\item{event_description}{Description}

\item{recorded_by}{User recording}
}
\value{
List with success status
}
\description{
Add Timeline Event
}
\keyword{internal}
