% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccg.R
\name{approve_ccg_version}
\alias{approve_ccg_version}
\title{Approve CCG Version}
\usage{
approve_ccg_version(version_id, approved_by)
}
\arguments{
\item{version_id}{Version ID}

\item{approved_by}{User approving}
}
\value{
List with success status
}
\description{
Approves a CCG version for use.
}
\keyword{internal}
