% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\name{approve_pia}
\alias{approve_pia}
\title{Approve PIA}
\usage{
approve_pia(assessment_id, approved_by)
}
\arguments{
\item{assessment_id}{Assessment ID}

\item{approved_by}{Approver name}
}
\value{
List with success status
}
\description{
Approve PIA
}
\keyword{internal}
