% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_module.R
\name{auth_ui}
\alias{auth_ui}
\title{Authentication Module UI}
\usage{
auth_ui(id)
}
\arguments{
\item{id}{The namespace id for the module}
}
\value{
A tagList containing the login modal UI
}
\description{
Authentication Module UI
}
\examples{
\dontrun{
# Embed the authentication module in a custom Shiny app.
# Most users will run the full application via
# `zzedc::launch_zzedc()`; the example below shows how to
# embed the module in a custom Shiny app instead.
if (interactive()) {
  ui <- shiny::fluidPage(
    auth_ui("module_id")
  )
  server <- function(input, output, session) {
    auth_server("module_id")
  }
  shiny::shinyApp(ui, server)
}
}
}
\keyword{internal}
