% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_dialect.R
\name{build_upsert}
\alias{build_upsert}
\title{Build UPSERT Statement}
\usage{
build_upsert(adapter, table_name, columns, conflict_columns, update_columns)
}
\arguments{
\item{adapter}{DatabaseAdapter instance}

\item{table_name}{Name of the table}

\item{columns}{Character vector of all columns}

\item{conflict_columns}{Character vector of columns that define uniqueness}

\item{update_columns}{Character vector of columns to update on conflict}
}
\value{
Character string with UPSERT SQL
}
\description{
Generates an INSERT ... ON CONFLICT / INSERT OR REPLACE statement
appropriate for the database backend.
}
\examples{
\dontrun{
  adapter <- create_db_adapter(config)
  sql <- build_upsert(adapter, "settings",
                      columns = c("key", "value", "updated_at"),
                      conflict_columns = c("key"),
                      update_columns = c("value", "updated_at"))
}

}
\keyword{internal}
