% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate Age}
\usage{
calculate_age(birth_date, reference_date = Sys.Date())
}
\arguments{
\item{birth_date}{Date of birth}

\item{reference_date}{Reference date (default: today)}
}
\value{
List with age
}
\description{
Calculate Age
}
\keyword{internal}
