% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{calculate_bsa}
\alias{calculate_bsa}
\title{Calculate Body Surface Area}
\usage{
calculate_bsa(weight, height, formula = "DUBOIS")
}
\arguments{
\item{weight}{Weight in kg}

\item{height}{Height in cm}

\item{formula}{Formula to use (DUBOIS, MOSTELLER)}
}
\value{
List with BSA
}
\description{
Calculate Body Surface Area
}
\keyword{internal}
