% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{calculate_date_diff}
\alias{calculate_date_diff}
\title{Calculate Date Difference}
\usage{
calculate_date_diff(start_date, end_date, unit = "days")
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}

\item{unit}{Unit (days, weeks, months, years)}
}
\value{
List with difference
}
\description{
Calculate Date Difference
}
\keyword{internal}
