% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_config.R
\name{clickhouse_config_template}
\alias{clickhouse_config_template}
\title{ClickHouse Configuration Template}
\usage{
clickhouse_config_template()
}
\value{
Character string with YAML template
}
\description{
ClickHouse Configuration Template
}
\keyword{internal}
