% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{close_db_pool}
\alias{close_db_pool}
\title{Close Connection Pool}
\usage{
close_db_pool(pool)
}
\arguments{
\item{pool}{Pool object}
}
\value{
NULL invisibly
}
\description{
Closes a database connection pool.
}
\keyword{internal}
