% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_version.R
\name{compare_crf_versions}
\alias{compare_crf_versions}
\title{Compare CRF Versions}
\usage{
compare_crf_versions(version_id_old, version_id_new, compared_by)
}
\arguments{
\item{version_id_old}{Old version ID}

\item{version_id_new}{New version ID}

\item{compared_by}{User performing comparison}
}
\value{
List with success status and comparison details
}
\description{
Compares two versions of a CRF.
}
\keyword{internal}
