% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_dialect.R
\name{concat_strings}
\alias{concat_strings}
\title{Build String Concatenation}
\usage{
concat_strings(adapter, ...)
}
\arguments{
\item{adapter}{DatabaseAdapter instance}

\item{...}{Column names or literals to concatenate}
}
\value{
Character string with concatenation expression
}
\description{
Returns the appropriate string concatenation expression for the backend.
}
\examples{
\dontrun{
  adapter <- create_db_adapter(config)
  expr <- concat_strings(adapter, "first_name", "' '", "last_name")
  # SQLite: first_name || ' ' || last_name
  # DuckDB/PostgreSQL: CONCAT(first_name, ' ', last_name)
}

}
\keyword{internal}
