% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_wizard_utils.R
\name{create_launch_script}
\alias{create_launch_script}
\title{Create Launch Script for New ZZedc Instance}
\usage{
create_launch_script(config_list, output_path)
}
\arguments{
\item{config_list}{Configuration from wizard}

\item{output_path}{Path where launch script will be written}
}
\value{
List with success status
}
\description{
Creates a customized launch script file that users can run to start the application
}
\keyword{internal}
