% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_templates.R
\name{crf_templates}
\alias{crf_templates}
\title{CRF Template Library System}
\description{
Provides pre-built CRF templates for common clinical trial forms.
Templates can be customized and used as starting points for new CRFs.
}
\keyword{internal}
