% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_validators.R
\name{days_between}
\alias{days_between}
\title{Calculate days between two dates}
\usage{
days_between(date1, date2)
}
\arguments{
\item{date1}{First date (Date object or string)}

\item{date2}{Second date (Date object or string)}
}
\value{
Number of days between dates (can be negative)
}
\description{
Returns the number of days between two dates.
}
\keyword{internal}
