% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{deactivate_condition}
\alias{deactivate_condition}
\title{Deactivate Condition}
\usage{
deactivate_condition(condition_id)
}
\arguments{
\item{condition_id}{Condition ID}
}
\value{
List with success status
}
\description{
Deactivate Condition
}
\keyword{internal}
