% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_library.R
\name{deactivate_library_field}
\alias{deactivate_library_field}
\title{Deactivate Library Field}
\usage{
deactivate_library_field(field_lib_id, deactivated_by)
}
\arguments{
\item{field_lib_id}{Field library ID}

\item{deactivated_by}{User deactivating}
}
\value{
List with success status
}
\description{
Deactivates a field (soft delete).
}
\keyword{internal}
