% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{disconnect_db}
\alias{disconnect_db}
\title{Disconnect from Database}
\usage{
disconnect_db(conn)
}
\arguments{
\item{conn}{DBI connection object}
}
\value{
NULL invisibly
}
\description{
Closes a database connection.
}
\keyword{internal}
