% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_module.R
\name{.deactivate_modal_dialog}
\alias{.deactivate_modal_dialog}
\title{Build the Deactivate-user modal body}
\usage{
.deactivate_modal_dialog(ns)
}
\description{
Build the Deactivate-user modal body
}
\keyword{internal}
