% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_module.R
\name{.reset_password_modal_dialog}
\alias{.reset_password_modal_dialog}
\title{Build the Reset-password modal body}
\usage{
.reset_password_modal_dialog(ns)
}
\description{
Build the Reset-password modal body
}
\keyword{internal}
