% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_module.R
\name{.user_modal_dialog}
\alias{.user_modal_dialog}
\title{Build the Add/Edit user modal body}
\usage{
.user_modal_dialog(
  ns,
  title,
  username = "",
  fullname = "",
  email = "",
  role = "Coordinator",
  active = TRUE
)
}
\description{
Build the Add/Edit user modal body
}
\keyword{internal}
