% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_config.R
\name{duckdb_config_template}
\alias{duckdb_config_template}
\title{DuckDB Configuration Template}
\usage{
duckdb_config_template()
}
\value{
Character string with YAML template
}
\description{
DuckDB Configuration Template
}
\keyword{internal}
