% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dsl_codegen.R
\name{evaluate_comparison_node}
\alias{evaluate_comparison_node}
\title{Evaluate comparison operation}
\usage{
evaluate_comparison_node(node, value, form_values, field_name)
}
\description{
Evaluate comparison operation
}
\keyword{internal}
