% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dsl_codegen.R
\name{evaluate_date_arithmetic_node}
\alias{evaluate_date_arithmetic_node}
\title{Evaluate date arithmetic operation}
\usage{
evaluate_date_arithmetic_node(node, value, form_values, field_name)
}
\description{
Evaluate date arithmetic operation
}
\keyword{internal}
