% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dsl_codegen.R
\name{evaluate_today_node}
\alias{evaluate_today_node}
\title{Evaluate today() function}
\usage{
evaluate_today_node(node, value, form_values, field_name)
}
\description{
Evaluate today() function
}
\keyword{internal}
