% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dsl_parser.R
\name{extract_ast_fields}
\alias{extract_ast_fields}
\title{Extract fields referenced in an AST}
\usage{
extract_ast_fields(ast)
}
\arguments{
\item{ast}{List representing the AST}
}
\value{
Character vector of field names
}
\description{
Returns all field names that are referenced in the validation rule.
}
\examples{
\dontrun{
ast <- parse_dsl_rule("x >= 18")
fields <- extract_ast_fields(ast)
}

}
\keyword{internal}
