% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{fail_controller_transfer}
\alias{fail_controller_transfer}
\title{Fail Controller Transfer}
\usage{
fail_controller_transfer(transfer_id, failure_reason, failed_by)
}
\arguments{
\item{transfer_id}{Transfer ID}

\item{failure_reason}{Reason for failure}

\item{failed_by}{User recording failure}
}
\value{
List with success status
}
\description{
Records failure of a controller-to-controller transfer.
}
\examples{
\dontrun{
fail_controller_transfer(
  transfer_id = 1,
  failure_reason = "Target controller API unavailable",
  failed_by = "dpo"
)
}
}
\keyword{internal}
