% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{generate_ae_number}
\alias{generate_ae_number}
\title{Generate AE Number}
\usage{
generate_ae_number(study_id, is_serious = FALSE)
}
\arguments{
\item{study_id}{Character: Study ID}

\item{is_serious}{Logical: Whether SAE}
}
\value{
Character: AE number
}
\description{
Generates a unique adverse event number.
}
\keyword{internal}
