% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erasure.R
\name{generate_erasure_number}
\alias{generate_erasure_number}
\title{Generate Erasure Request Number}
\usage{
generate_erasure_number()
}
\value{
Character: Request number
}
\description{
Generate Erasure Request Number
}
\keyword{internal}
