% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erasure.R
\name{generate_hold_number}
\alias{generate_hold_number}
\title{Generate Legal Hold Number}
\usage{
generate_hold_number()
}
\value{
Character: Hold number
}
\description{
Generate Legal Hold Number
}
\keyword{internal}
