% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{generate_objection_number}
\alias{generate_objection_number}
\title{Generate Objection Request Number}
\usage{
generate_objection_number()
}
\value{
Character string in format OBJ-TIMESTAMP-RANDOM
}
\description{
Generate Objection Request Number
}
\keyword{internal}
