% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectification.R
\name{generate_rect_number}
\alias{generate_rect_number}
\title{Generate Rectification Request Number}
\usage{
generate_rect_number()
}
\value{
Character: Request number
}
\description{
Generates a unique rectification request number.
}
\keyword{internal}
