% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsar.R
\name{generate_request_number}
\alias{generate_request_number}
\title{Generate Request Number}
\usage{
generate_request_number(request_type)
}
\arguments{
\item{request_type}{Character: Request type}
}
\value{
Character: Request number
}
\description{
Generates a unique DSAR request number.
}
\keyword{internal}
