% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.R
\name{generate_restriction_number}
\alias{generate_restriction_number}
\title{Generate Restriction Request Number}
\usage{
generate_restriction_number()
}
\value{
Character string in format RESTR-TIMESTAMP-RANDOM
}
\description{
Generates a unique restriction request number.
}
\keyword{internal}
