% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{generate_security_salt}
\alias{generate_security_salt}
\title{Generate Security Salt}
\usage{
generate_security_salt()
}
\value{
Character string of random salt
}
\description{
Creates a random 32-character salt for password hashing
}
\keyword{internal}
