% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electronic_signatures.R
\name{generate_signature_code}
\alias{generate_signature_code}
\title{Generate Signature Code}
\usage{
generate_signature_code()
}
\value{
Character: Unique signature code
}
\description{
Generates a unique signature code.
}
\keyword{internal}
