% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dsl_sql_codegen.R
\name{generate_sql_if}
\alias{generate_sql_if}
\title{Generate SQL for if/then/else logic}
\usage{
generate_sql_if(ast, field_name)
}
\description{
Generate SQL for if/then/else logic
}
\keyword{internal}
