% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dsl_sql_codegen.R
\name{generate_sql_where_clause}
\alias{generate_sql_where_clause}
\title{Generate WHERE clause from AST}
\usage{
generate_sql_where_clause(ast, field_name = NULL)
}
\arguments{
\item{ast}{Abstract Syntax Tree}

\item{field_name}{Field name for simple comparisons}
}
\value{
SQL WHERE clause (without "WHERE" keyword)
}
\description{
Converts AST into a SQL WHERE clause that identifies violations.
}
\keyword{internal}
