% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{get_ae_severity_grades}
\alias{get_ae_severity_grades}
\title{Get AE Severity Grades}
\usage{
get_ae_severity_grades()
}
\value{
Named list of severity grades
}
\description{
Returns valid severity grades with descriptions.
}
\keyword{internal}
