% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dsl_parser.R
\name{get_allow_values}
\alias{get_allow_values}
\title{Get allowed values from an allow rule}
\usage{
get_allow_values(ast)
}
\arguments{
\item{ast}{List representing the AST with type "allow"}
}
\value{
List of allowed values
}
\description{
Get allowed values from an allow rule
}
\keyword{internal}
