% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\name{get_breach_incident}
\alias{get_breach_incident}
\title{Get Breach Incident}
\usage{
get_breach_incident(incident_id)
}
\arguments{
\item{incident_id}{Incident ID}
}
\value{
List with incident details
}
\description{
Get Breach Incident
}
\keyword{internal}
