% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\name{get_breach_severities}
\alias{get_breach_severities}
\title{Get Breach Severities}
\usage{
get_breach_severities()
}
\value{
Named character vector
}
\description{
Get Breach Severities
}
\keyword{internal}
