% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\name{get_breach_statistics}
\alias{get_breach_statistics}
\title{Get Breach Statistics}
\usage{
get_breach_statistics()
}
\value{
List with statistics
}
\description{
Get Breach Statistics
}
\keyword{internal}
