% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\name{get_breach_statuses}
\alias{get_breach_statuses}
\title{Get Breach Statuses}
\usage{
get_breach_statuses()
}
\value{
Named character vector
}
\description{
Get Breach Statuses
}
\keyword{internal}
