% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{get_calculated_fields}
\alias{get_calculated_fields}
\title{Get Calculated Fields}
\usage{
get_calculated_fields(form_id = NULL, include_inactive = FALSE)
}
\arguments{
\item{form_id}{Optional form filter}

\item{include_inactive}{Include inactive}
}
\value{
List with calculations
}
\description{
Get Calculated Fields
}
\keyword{internal}
