% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{get_calculation_history}
\alias{get_calculation_history}
\title{Get Calculation History}
\usage{
get_calculation_history(calc_id, subject_id = NULL)
}
\arguments{
\item{calc_id}{Calculation ID}

\item{subject_id}{Optional subject filter}
}
\value{
List with history
}
\description{
Get Calculation History
}
\keyword{internal}
