% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{get_calculation_types}
\alias{get_calculation_types}
\title{Get Calculation Types}
\usage{
get_calculation_types()
}
\value{
Named character vector
}
\description{
Get Calculation Types
}
\keyword{internal}
