% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccg.R
\name{get_ccg_fields}
\alias{get_ccg_fields}
\title{Get CCG Fields}
\usage{
get_ccg_fields(form_id, section_name = NULL)
}
\arguments{
\item{form_id}{Form ID}

\item{section_name}{Optional filter by section}
}
\value{
List with success status and fields
}
\description{
Retrieves fields for a CCG form.
}
\keyword{internal}
