% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccg.R
\name{get_ccg_form_categories}
\alias{get_ccg_form_categories}
\title{Get Form Categories}
\usage{
get_ccg_form_categories()
}
\value{
Named character vector of form categories
}
\description{
Get Form Categories
}
\keyword{internal}
