% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccg.R
\name{get_ccg_form_statuses}
\alias{get_ccg_form_statuses}
\title{Get Form Statuses}
\usage{
get_ccg_form_statuses()
}
\value{
Named character vector of form statuses
}
\description{
Get Form Statuses
}
\keyword{internal}
