% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_library.R
\name{get_cdisc_domains}
\alias{get_cdisc_domains}
\title{Get CDISC Domains}
\usage{
get_cdisc_domains()
}
\value{
Named character vector of CDISC domains
}
\description{
Get CDISC Domains
}
\keyword{internal}
