% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{get_change_approvals}
\alias{get_change_approvals}
\title{Get Change Approvals}
\usage{
get_change_approvals(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with approvals
}
\description{
Get Change Approvals
}
\keyword{internal}
