% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{get_change_categories}
\alias{get_change_categories}
\title{Get Change Categories}
\usage{
get_change_categories()
}
\value{
Named character vector
}
\description{
Get Change Categories
}
\keyword{internal}
