% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{get_change_control_statistics}
\alias{get_change_control_statistics}
\title{Get Change Control Statistics}
\usage{
get_change_control_statistics()
}
\value{
List with statistics
}
\description{
Get Change Control Statistics
}
\keyword{internal}
