% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{get_change_priorities}
\alias{get_change_priorities}
\title{Get Change Priorities}
\usage{
get_change_priorities()
}
\value{
Named character vector
}
\description{
Get Change Priorities
}
\keyword{internal}
