% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{get_change_request_details}
\alias{get_change_request_details}
\title{Get Change Request Details}
\usage{
get_change_request_details(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with full request details
}
\description{
Get Change Request Details
}
\keyword{internal}
