% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{get_change_requests}
\alias{get_change_requests}
\title{Get Change Requests}
\usage{
get_change_requests(status = NULL, request_type = NULL, change_category = NULL)
}
\arguments{
\item{status}{Optional status filter}

\item{request_type}{Optional type filter}

\item{change_category}{Optional category filter}
}
\value{
List with change requests
}
\description{
Get Change Requests
}
\keyword{internal}
