% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{get_checklist_categories}
\alias{get_checklist_categories}
\title{Get Checklist Categories}
\usage{
get_checklist_categories()
}
\value{
Named character vector
}
\description{
Get Checklist Categories
}
\keyword{internal}
