% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{get_checklist_status}
\alias{get_checklist_status}
\title{Get Checklist Status}
\usage{
get_checklist_status(closeout_id)
}
\arguments{
\item{closeout_id}{Closeout ID}
}
\value{
List with checklist
}
\description{
Get Checklist Status
}
\keyword{internal}
