% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{get_closeout_statistics}
\alias{get_closeout_statistics}
\title{Get Closeout Statistics}
\usage{
get_closeout_statistics()
}
\value{
List with statistics
}
\description{
Get Closeout Statistics
}
\keyword{internal}
