% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{get_closeout_statuses}
\alias{get_closeout_statuses}
\title{Get Closeout Status Values}
\usage{
get_closeout_statuses()
}
\value{
Named character vector
}
\description{
Get Closeout Status Values
}
\keyword{internal}
