% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{get_condition_operators}
\alias{get_condition_operators}
\title{Get Operators}
\usage{
get_condition_operators()
}
\value{
Named character vector
}
\description{
Get Operators
}
\keyword{internal}
