% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{get_condition_types}
\alias{get_condition_types}
\title{Get Condition Types}
\usage{
get_condition_types()
}
\value{
Named character vector
}
\description{
Get Condition Types
}
\keyword{internal}
