% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{get_consent_legal_bases}
\alias{get_consent_legal_bases}
\title{Get Consent Legal Bases}
\usage{
get_consent_legal_bases()
}
\value{
Named character vector of legal bases
}
\description{
Get Consent Legal Bases
}
\keyword{internal}
