% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{get_consent_methods}
\alias{get_consent_methods}
\title{Get Consent Methods}
\usage{
get_consent_methods()
}
\value{
Named character vector of consent methods
}
\description{
Get Consent Methods
}
\keyword{internal}
