% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{get_consent_statuses}
\alias{get_consent_statuses}
\title{Get Consent Statuses}
\usage{
get_consent_statuses()
}
\value{
Named character vector of consent statuses
}
\description{
Get Consent Statuses
}
\keyword{internal}
