% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_version.R
\name{get_crf_change_types}
\alias{get_crf_change_types}
\title{Get Change Types}
\usage{
get_crf_change_types()
}
\value{
Named character vector of change types
}
\description{
Get Change Types
}
\keyword{internal}
