% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_version.R
\name{get_crf_version_statuses}
\alias{get_crf_version_statuses}
\title{Get Version Statuses}
\usage{
get_crf_version_statuses()
}
\value{
Named character vector of version statuses
}
\description{
Get Version Statuses
}
\keyword{internal}
