% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_version.R
\name{get_crf_version_types}
\alias{get_crf_version_types}
\title{Get Version Types}
\usage{
get_crf_version_types()
}
\value{
Named character vector of version types
}
\description{
Get Version Types
}
\keyword{internal}
