% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{get_db_dialect}
\alias{get_db_dialect}
\title{Get SQL Dialect}
\usage{
get_db_dialect()
}
\value{
List with dialect-specific SQL fragments
}
\description{
Returns dialect information for the current backend.
}
\keyword{internal}
