% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{get_dependency_types}
\alias{get_dependency_types}
\title{Get Dependency Types}
\usage{
get_dependency_types()
}
\value{
Named character vector
}
\description{
Get Dependency Types
}
\keyword{internal}
