% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{get_dependent_fields}
\alias{get_dependent_fields}
\title{Get Dependent Fields}
\usage{
get_dependent_fields(parent_field_code)
}
\arguments{
\item{parent_field_code}{Parent field code}
}
\value{
List with child fields
}
\description{
Get Dependent Fields
}
\keyword{internal}
