% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{get_designer_field_types}
\alias{get_designer_field_types}
\title{Get Field Types}
\usage{
get_designer_field_types()
}
\value{
Named character vector
}
\description{
Get Field Types
}
\keyword{internal}
