% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{get_designer_forms}
\alias{get_designer_forms}
\title{Get Designer Forms List}
\usage{
get_designer_forms(status = NULL)
}
\arguments{
\item{status}{Optional status filter}
}
\value{
List with forms
}
\description{
Get Designer Forms List
}
\keyword{internal}
