% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsar.R
\name{get_dsar_collected_data}
\alias{get_dsar_collected_data}
\title{Get Collected Data}
\usage{
get_dsar_collected_data(request_id, db_path = NULL)
}
\arguments{
\item{request_id}{Integer: Request ID}

\item{db_path}{Character: Database path (optional)}
}
\value{
Data frame with collected data
}
\description{
Retrieves collected data records for a DSAR request.
}
\keyword{internal}
