% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsar.R
\name{get_dsar_statistics}
\alias{get_dsar_statistics}
\title{Get DSAR Statistics}
\usage{
get_dsar_statistics(db_path = NULL)
}
\arguments{
\item{db_path}{Character: Database path (optional)}
}
\value{
List with DSAR statistics
}
\description{
Returns DSAR statistics.
}
\keyword{internal}
