% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsar.R
\name{get_dsar_statuses}
\alias{get_dsar_statuses}
\title{Get DSAR Status Values}
\usage{
get_dsar_statuses()
}
\value{
Named list of status values
}
\description{
Returns valid DSAR status values with descriptions.
}
\keyword{internal}
