% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_gsheets_integration.R
\name{get_dsl_rule_categories}
\alias{get_dsl_rule_categories}
\title{Rule categories for validation}
\usage{
get_dsl_rule_categories()
}
\description{
Rule categories for validation
}
\keyword{internal}
