% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_gsheets_integration.R
\name{get_dsl_severity_levels}
\alias{get_dsl_severity_levels}
\title{Severity levels for validation rules}
\usage{
get_dsl_severity_levels()
}
\description{
Severity levels for validation rules
}
\keyword{internal}
