% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erasure.R
\name{get_erasure_methods}
\alias{get_erasure_methods}
\title{Get Erasure Methods}
\usage{
get_erasure_methods()
}
\value{
Named list of erasure methods
}
\description{
Returns valid erasure methods.
}
\keyword{internal}
