% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erasure.R
\name{get_erasure_statuses}
\alias{get_erasure_statuses}
\title{Get Erasure Statuses}
\usage{
get_erasure_statuses()
}
\value{
Named list of status values
}
\description{
Returns valid erasure request statuses.
}
\keyword{internal}
