% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{get_field_conditions}
\alias{get_field_conditions}
\title{Get Field Conditions}
\usage{
get_field_conditions(
  form_id = NULL,
  target_field_code = NULL,
  include_inactive = FALSE
)
}
\arguments{
\item{form_id}{Optional form filter}

\item{target_field_code}{Optional target field filter}

\item{include_inactive}{Include inactive conditions}
}
\value{
List with conditions
}
\description{
Get Field Conditions
}
\keyword{internal}
