% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_library.R
\name{get_field_data_types}
\alias{get_field_data_types}
\title{Get Field Data Types}
\usage{
get_field_data_types()
}
\value{
Named character vector of data types
}
\description{
Get Field Data Types
}
\keyword{internal}
