% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_library.R
\name{get_field_library_categories}
\alias{get_field_library_categories}
\title{Get Field Categories}
\usage{
get_field_library_categories()
}
\value{
Named character vector of field categories
}
\description{
Get Field Categories
}
\keyword{internal}
