% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\name{get_gdpr_legal_bases}
\alias{get_gdpr_legal_bases}
\title{Get Legal Bases (GDPR Article 6)}
\usage{
get_gdpr_legal_bases()
}
\value{
Named character vector
}
\description{
Get Legal Bases (GDPR Article 6)
}
\keyword{internal}
