% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{get_implementation_status}
\alias{get_implementation_status}
\title{Get Implementation Status}
\usage{
get_implementation_status(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with implementation steps
}
\description{
Get Implementation Status
}
\keyword{internal}
