% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erasure.R
\name{get_legal_hold_types}
\alias{get_legal_hold_types}
\title{Get Legal Hold Types}
\usage{
get_legal_hold_types()
}
\value{
Named list of hold types
}
\description{
Returns valid legal hold types.
}
\keyword{internal}
