% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_library.R
\name{get_library_field}
\alias{get_library_field}
\title{Get Library Field by Code}
\usage{
get_library_field(field_code)
}
\arguments{
\item{field_code}{Field code}
}
\value{
List with success status and field details
}
\description{
Retrieves a single field by code.
}
\keyword{internal}
