% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{get_marketing_channels}
\alias{get_marketing_channels}
\title{Get Marketing Channels}
\usage{
get_marketing_channels()
}
\value{
Named character vector of marketing channels
}
\description{
Returns valid marketing channel types.
}
\keyword{internal}
