% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{get_objection_activities}
\alias{get_objection_activities}
\title{Get Objection Activities}
\usage{
get_objection_activities(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with success status and activities
}
\description{
Get Objection Activities
}
\keyword{internal}
