% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{get_objection_history}
\alias{get_objection_history}
\title{Get Objection History}
\usage{
get_objection_history(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with success status and history
}
\description{
Get Objection History
}
\keyword{internal}
