% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{get_objection_request}
\alias{get_objection_request}
\title{Get Objection Request}
\usage{
get_objection_request(request_id = NULL, request_number = NULL)
}
\arguments{
\item{request_id}{Optional request ID}

\item{request_number}{Optional request number}
}
\value{
List with success status and request data
}
\description{
Get Objection Request
}
\keyword{internal}
