% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{get_objection_statistics}
\alias{get_objection_statistics}
\title{Get Objection Statistics}
\usage{
get_objection_statistics()
}
\value{
List with statistics
}
\description{
Get Objection Statistics
}
\keyword{internal}
