% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{get_objection_types}
\alias{get_objection_types}
\title{Get Objection Types}
\usage{
get_objection_types()
}
\value{
Named character vector of objection types
}
\description{
Returns valid objection types under Article 21.
}
\keyword{internal}
