% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{get_pending_withdrawal_requests}
\alias{get_pending_withdrawal_requests}
\title{Get Pending Withdrawal Requests}
\usage{
get_pending_withdrawal_requests()
}
\value{
List with success status and requests
}
\description{
Gets all pending withdrawal requests.
}
\keyword{internal}
