% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\name{get_pia_risk_categories}
\alias{get_pia_risk_categories}
\title{Get Risk Categories}
\usage{
get_pia_risk_categories()
}
\value{
Named character vector
}
\description{
Get Risk Categories
}
\keyword{internal}
