% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{get_portability_exceptions}
\alias{get_portability_exceptions}
\title{Get Portability Exceptions}
\usage{
get_portability_exceptions()
}
\value{
Named character vector of exception grounds
}
\description{
Returns valid exception grounds under Article 20(4).
}
\examples{
exceptions <- get_portability_exceptions()
}
\keyword{internal}
