% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{get_portability_exports}
\alias{get_portability_exports}
\title{Get Portability Exports}
\usage{
get_portability_exports(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with success status and exports
}
\description{
Retrieves exports for a portability request.
}
\examples{
\dontrun{
exports <- get_portability_exports(request_id = 1)
}
}
\keyword{internal}
