% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{get_portability_request_types}
\alias{get_portability_request_types}
\title{Get Portability Request Types}
\usage{
get_portability_request_types()
}
\value{
Named character vector of request types
}
\description{
Returns valid portability request types.
}
\examples{
types <- get_portability_request_types()
}
\keyword{internal}
