% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_linkage.R
\name{get_protocol_crf_links}
\alias{get_protocol_crf_links}
\title{Get Protocol CRF Links}
\usage{
get_protocol_crf_links(protocol_id)
}
\arguments{
\item{protocol_id}{Protocol ID}
}
\value{
List with links
}
\description{
Get Protocol CRF Links
}
\keyword{internal}
