% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_linkage.R
\name{get_protocol_definitions}
\alias{get_protocol_definitions}
\title{Get Protocol Definitions}
\usage{
get_protocol_definitions(include_inactive = FALSE)
}
\arguments{
\item{include_inactive}{Include inactive protocols}
}
\value{
List with protocols
}
\description{
Get Protocol Definitions
}
\keyword{internal}
