% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_linkage.R
\name{get_protocol_linkage_statistics}
\alias{get_protocol_linkage_statistics}
\title{Get Protocol Linkage Statistics}
\usage{
get_protocol_linkage_statistics()
}
\value{
List with statistics
}
\description{
Get Protocol Linkage Statistics
}
\keyword{internal}
