% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{get_reconciliation_status}
\alias{get_reconciliation_status}
\title{Get Reconciliation Status}
\usage{
get_reconciliation_status(closeout_id)
}
\arguments{
\item{closeout_id}{Closeout ID}
}
\value{
List with reconciliations
}
\description{
Get Reconciliation Status
}
\keyword{internal}
