% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectification.R
\name{get_rectification_statuses}
\alias{get_rectification_statuses}
\title{Get Rectification Statuses}
\usage{
get_rectification_statuses()
}
\value{
Named list of status values
}
\description{
Returns valid rectification request statuses.
}
\keyword{internal}
