% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.R
\name{get_restriction_third_parties}
\alias{get_restriction_third_parties}
\title{Get Restriction Third Parties}
\usage{
get_restriction_third_parties(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with success status and recipients
}
\description{
Retrieves third-party recipients for a restriction request.
}
\examples{
\dontrun{
recipients <- get_restriction_third_parties(request_id = 1)
}
}
\keyword{internal}
