% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{get_retention_actions}
\alias{get_retention_actions}
\title{Get Retention Actions}
\usage{
get_retention_actions()
}
\value{
Named character vector of actions on expiry
}
\description{
Get Retention Actions
}
\keyword{internal}
