% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{get_retention_bases}
\alias{get_retention_bases}
\title{Get Retention Bases}
\usage{
get_retention_bases()
}
\value{
Named character vector of retention legal bases
}
\description{
Get Retention Bases
}
\keyword{internal}
