% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_review.R
\name{get_review_decisions}
\alias{get_review_decisions}
\title{Get Review Decisions}
\usage{
get_review_decisions()
}
\value{
Named character vector of decisions
}
\description{
Get Review Decisions
}
\keyword{internal}
