% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\name{get_risk_assessments}
\alias{get_risk_assessments}
\title{Get Risk Assessments}
\usage{
get_risk_assessments(assessment_id)
}
\arguments{
\item{assessment_id}{Assessment ID}
}
\value{
List with risks
}
\description{
Get Risk Assessments
}
\keyword{internal}
