% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electronic_signatures.R
\name{get_signature_meanings}
\alias{get_signature_meanings}
\title{Get Signature Meanings}
\usage{
get_signature_meanings()
}
\value{
Named list of signature meanings
}
\description{
Returns valid signature meaning codes with descriptions.
}
\keyword{internal}
