% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_review.R
\name{get_stage_reviewers}
\alias{get_stage_reviewers}
\title{Get Stage Reviewers}
\usage{
get_stage_reviewers(stage_id)
}
\arguments{
\item{stage_id}{Stage ID}
}
\value{
List with success status and reviewers
}
\description{
Retrieves reviewers for a stage.
}
\keyword{internal}
