% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{get_standard_formulas}
\alias{get_standard_formulas}
\title{Get Standard Formulas}
\usage{
get_standard_formulas(category = NULL)
}
\arguments{
\item{category}{Optional category filter}
}
\value{
List with formulas
}
\description{
Get Standard Formulas
}
\keyword{internal}
