% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_templates.R
\name{get_template_categories}
\alias{get_template_categories}
\title{Get Template Categories}
\usage{
get_template_categories()
}
\value{
Named character vector
}
\description{
Get Template Categories
}
\keyword{internal}
