% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_templates.R
\name{get_template_fields}
\alias{get_template_fields}
\title{Get Template Fields}
\usage{
get_template_fields(template_id)
}
\arguments{
\item{template_id}{Template ID}
}
\value{
List with fields
}
\description{
Get Template Fields
}
\keyword{internal}
