% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_templates.R
\name{get_template_statistics}
\alias{get_template_statistics}
\title{Get Template Statistics}
\usage{
get_template_statistics()
}
\value{
List with statistics
}
\description{
Get Template Statistics
}
\keyword{internal}
