% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_gsheets_integration.R
\name{get_user_dsl_permissions}
\alias{get_user_dsl_permissions}
\title{Get User DSL Permissions}
\usage{
get_user_dsl_permissions(user_role)
}
\arguments{
\item{user_role}{User's role}
}
\value{
List of permissions
}
\description{
Returns all DSL rule permissions for a user role.
}
\keyword{internal}
