% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{get_validation_rule_categories}
\alias{get_validation_rule_categories}
\title{Get Rule Categories}
\usage{
get_validation_rule_categories()
}
\value{
Named character vector
}
\description{
Get Rule Categories
}
\keyword{internal}
