% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{get_validation_rule_types}
\alias{get_validation_rule_types}
\title{Get Rule Types}
\usage{
get_validation_rule_types()
}
\value{
Named character vector
}
\description{
Get Rule Types
}
\keyword{internal}
