% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{get_validation_severities}
\alias{get_validation_severities}
\title{Get Severity Levels}
\usage{
get_validation_severities()
}
\value{
Named character vector
}
\description{
Get Severity Levels
}
\keyword{internal}
