% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{get_validation_statistics}
\alias{get_validation_statistics}
\title{Get Validation Statistics}
\usage{
get_validation_statistics()
}
\value{
List with statistics
}
\description{
Get Validation Statistics
}
\keyword{internal}
