% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_import_module.R
\name{instrument_import_server}
\alias{instrument_import_server}
\title{Instrument Import Module Server}
\usage{
instrument_import_server(id, db_conn, instruments_dir = "instruments/")
}
\arguments{
\item{id}{Namespace ID for the module}

\item{db_conn}{Reactive database connection}

\item{instruments_dir}{Directory containing instrument CSV files}
}
\description{
Instrument Import Module Server
}
\keyword{internal}
