% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_import_module.R
\name{instrument_import_ui}
\alias{instrument_import_ui}
\title{Instrument Import Module UI}
\usage{
instrument_import_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
A tagList containing the instrument import UI
}
\description{
Instrument Import Module UI
}
\keyword{internal}
