% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{is_field_visible}
\alias{is_field_visible}
\title{Check if field is visible}
\usage{
is_field_visible(field_name, field_config, form_values)
}
\arguments{
\item{field_name}{Name of field to check}

\item{field_config}{Field configuration with branching rules}

\item{form_values}{Named list of current form values}
}
\value{
Logical, TRUE if field should be visible
}
\description{
Determines if a field should be displayed based on current form values.
}
\keyword{internal}
