% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{load_standard_checklist}
\alias{load_standard_checklist}
\title{Load Standard Checklist}
\usage{
load_standard_checklist(closeout_id)
}
\arguments{
\item{closeout_id}{Closeout ID}
}
\value{
List with success status
}
\description{
Load Standard Checklist
}
\keyword{internal}
