% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{load_standard_formulas}
\alias{load_standard_formulas}
\title{Load Standard Formulas}
\usage{
load_standard_formulas()
}
\value{
List with success status
}
\description{
Load Standard Formulas
}
\keyword{internal}
