% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_templates.R
\name{load_standard_templates}
\alias{load_standard_templates}
\title{Load Standard Templates}
\usage{
load_standard_templates(loaded_by)
}
\arguments{
\item{loaded_by}{User loading templates}
}
\value{
List with count
}
\description{
Load Standard Templates
}
\keyword{internal}
