% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{log_consent_action}
\alias{log_consent_action}
\title{Log Consent Action}
\usage{
log_consent_action(
  consent_id,
  action,
  action_details,
  previous_state,
  new_state,
  performed_by
)
}
\description{
Log Consent Action
}
\keyword{internal}
