% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{log_objection_action}
\alias{log_objection_action}
\title{Log Objection Action}
\usage{
log_objection_action(
  request_id,
  action,
  action_details,
  performed_by,
  activity_id = NULL
)
}
\description{
Log Objection Action
}
\keyword{internal}
