% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling_wrappers.R
\name{make_profiled}
\alias{make_profiled}
\title{Create Profiled Function Wrapper}
\usage{
make_profiled(fn, category, operation)
}
\arguments{
\item{fn}{Function to wrap}

\item{category}{Profiling category}

\item{operation}{Operation name}
}
\value{
Wrapped function with profiling
}
\description{
Factory function to create a profiled version of any function.
}
\examples{
\dontrun{
  profiled_mean <- make_profiled(mean, "stats", "mean")
  result <- profiled_mean(1:1000000)
}

}
\keyword{internal}
