% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\name{notify_data_subjects}
\alias{notify_data_subjects}
\title{Notify Data Subjects}
\usage{
notify_data_subjects(
  incident_id,
  notified_by,
  subjects_notified,
  notification_method = NULL,
  notification_content = NULL
)
}
\arguments{
\item{incident_id}{Incident ID}

\item{notified_by}{User notifying}

\item{subjects_notified}{Number notified}

\item{notification_method}{Method used}

\item{notification_content}{Content sent}
}
\value{
List with success status
}
\description{
Notify Data Subjects
}
\keyword{internal}
