% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\name{notify_dpo}
\alias{notify_dpo}
\title{Notify DPO}
\usage{
notify_dpo(incident_id, notified_by)
}
\arguments{
\item{incident_id}{Incident ID}

\item{notified_by}{User notifying}
}
\value{
List with success status
}
\description{
Notify DPO
}
\keyword{internal}
