% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_dsl_parser.R
\name{Parser}
\alias{Parser}
\title{Parser class}
\description{
Parser class

Parser class
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Parser-new}{\code{Parser$new()}}
\item \href{#method-Parser-parse}{\code{Parser$parse()}}
\item \href{#method-Parser-parse_expression}{\code{Parser$parse_expression()}}
\item \href{#method-Parser-parse_or_expression}{\code{Parser$parse_or_expression()}}
\item \href{#method-Parser-parse_and_expression}{\code{Parser$parse_and_expression()}}
\item \href{#method-Parser-parse_comparison}{\code{Parser$parse_comparison()}}
\item \href{#method-Parser-parse_primary_comparison}{\code{Parser$parse_primary_comparison()}}
\item \href{#method-Parser-parse_in_without_field}{\code{Parser$parse_in_without_field()}}
\item \href{#method-Parser-parse_notin_without_field}{\code{Parser$parse_notin_without_field()}}
\item \href{#method-Parser-parse_between}{\code{Parser$parse_between()}}
\item \href{#method-Parser-parse_between_after_field}{\code{Parser$parse_between_after_field()}}
\item \href{#method-Parser-parse_in_after_field}{\code{Parser$parse_in_after_field()}}
\item \href{#method-Parser-parse_notin_after_field}{\code{Parser$parse_notin_after_field()}}
\item \href{#method-Parser-parse_if_expression}{\code{Parser$parse_if_expression()}}
\item \href{#method-Parser-parse_required}{\code{Parser$parse_required()}}
\item \href{#method-Parser-parse_allow}{\code{Parser$parse_allow()}}
\item \href{#method-Parser-parse_term}{\code{Parser$parse_term()}}
\item \href{#method-Parser-peek}{\code{Parser$peek()}}
\item \href{#method-Parser-consume}{\code{Parser$consume()}}
\item \href{#method-Parser-clone}{\code{Parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-new"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$new(tokens)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse}{}}}
\subsection{Method \code{parse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_expression"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_expression}{}}}
\subsection{Method \code{parse_expression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_expression()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_or_expression"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_or_expression}{}}}
\subsection{Method \code{parse_or_expression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_or_expression()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_and_expression"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_and_expression}{}}}
\subsection{Method \code{parse_and_expression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_and_expression()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_comparison"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_comparison}{}}}
\subsection{Method \code{parse_comparison()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_comparison()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_primary_comparison"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_primary_comparison}{}}}
\subsection{Method \code{parse_primary_comparison()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_primary_comparison()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_in_without_field"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_in_without_field}{}}}
\subsection{Method \code{parse_in_without_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_in_without_field()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_notin_without_field"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_notin_without_field}{}}}
\subsection{Method \code{parse_notin_without_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_notin_without_field()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_between"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_between}{}}}
\subsection{Method \code{parse_between()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_between()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_between_after_field"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_between_after_field}{}}}
\subsection{Method \code{parse_between_after_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_between_after_field(field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_in_after_field"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_in_after_field}{}}}
\subsection{Method \code{parse_in_after_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_in_after_field(field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_notin_after_field"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_notin_after_field}{}}}
\subsection{Method \code{parse_notin_after_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_notin_after_field(field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_if_expression"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_if_expression}{}}}
\subsection{Method \code{parse_if_expression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_if_expression()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_required"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_required}{}}}
\subsection{Method \code{parse_required()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_required()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_allow"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_allow}{}}}
\subsection{Method \code{parse_allow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_allow()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_term"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_term}{}}}
\subsection{Method \code{parse_term()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_term()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-peek"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-peek}{}}}
\subsection{Method \code{peek()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$peek()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-consume"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-consume}{}}}
\subsection{Method \code{consume()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$consume(expected_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
