% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_config.R
\name{postgresql_config_template}
\alias{postgresql_config_template}
\title{PostgreSQL Configuration Template}
\usage{
postgresql_config_template()
}
\value{
Character string with YAML template
}
\description{
PostgreSQL Configuration Template
}
\keyword{internal}
