% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling_wrappers.R
\name{profiled_fread}
\alias{profiled_fread}
\title{Profiled Data Table Read}
\usage{
profiled_fread(input, ...)
}
\arguments{
\item{input}{File path or URL}

\item{...}{Additional arguments passed to data.table::fread}
}
\value{
data.table object
}
\description{
Wrapper around data.table::fread with profiling instrumentation.
}
\keyword{internal}
