% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling_wrappers.R
\name{profiled_fwrite}
\alias{profiled_fwrite}
\title{Profiled Data Table Write}
\usage{
profiled_fwrite(x, file, ...)
}
\arguments{
\item{x}{data.table or data.frame to write}

\item{file}{Output file path}

\item{...}{Additional arguments passed to data.table::fwrite}
}
\value{
Invisible NULL
}
\description{
Wrapper around data.table::fwrite with profiling instrumentation.
}
\keyword{internal}
