% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling_wrappers.R
\name{profiled_http}
\alias{profiled_http}
\title{Profiled HTTP Request}
\usage{
profiled_http(method, url, ...)
}
\arguments{
\item{method}{HTTP method ("GET", "POST", etc.)}

\item{url}{Request URL}

\item{...}{Additional arguments passed to httr verb function}
}
\value{
httr response object
}
\description{
Wrapper around httr requests with profiling instrumentation.
}
\keyword{internal}
