% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling_wrappers.R
\name{profiled_json_parse}
\alias{profiled_json_parse}
\title{Profiled JSON Parse}
\usage{
profiled_json_parse(txt, ...)
}
\arguments{
\item{txt}{JSON string or file path}

\item{...}{Additional arguments passed to jsonlite::fromJSON}
}
\value{
Parsed R object
}
\description{
Wrapper around jsonlite::fromJSON with profiling instrumentation.
}
\examples{
\dontrun{
  profiling_enable()
  data <- profiled_json_parse('{"name": "test", "value": 123}')
  profiling_report()
}

}
\keyword{internal}
