% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling_wrappers.R
\name{profiled_json_serialize}
\alias{profiled_json_serialize}
\title{Profiled JSON Serialize}
\usage{
profiled_json_serialize(x, ...)
}
\arguments{
\item{x}{R object to serialize}

\item{...}{Additional arguments passed to jsonlite::toJSON}
}
\value{
JSON string
}
\description{
Wrapper around jsonlite::toJSON with profiling instrumentation.
}
\keyword{internal}
