% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{profiling_data}
\alias{profiling_data}
\title{Get Raw Profiling Data}
\usage{
profiling_data()
}
\value{
List of profiling event records
}
\description{
Returns the raw list of profiling events.
}
\keyword{internal}
