% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{profiling_disable}
\alias{profiling_disable}
\title{Disable Profiling}
\usage{
profiling_disable()
}
\value{
Invisible FALSE
}
\description{
Disables profiling. Collected data is retained until next enable.
}
\examples{
# Turn the profiling subsystem off (e.g., before deploying
# to production).
profiling_disable()
}
