% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{profiling_is_enabled}
\alias{profiling_is_enabled}
\title{Check if Profiling is Enabled}
\usage{
profiling_is_enabled()
}
\value{
Logical TRUE if profiling is currently enabled
}
\description{
Check if Profiling is Enabled
}
\keyword{internal}
