% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_dashboard_module.R
\name{quality_dashboard_ui}
\alias{quality_dashboard_ui}
\title{Quality Dashboard Module UI}
\usage{
quality_dashboard_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
A tagList containing the quality dashboard UI
}
\description{
Quality Dashboard Module UI
}
\keyword{internal}
