% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{record_data_reconciliation}
\alias{record_data_reconciliation}
\title{Record Data Reconciliation}
\usage{
record_data_reconciliation(
  closeout_id,
  reconciliation_type,
  performed_by,
  source_system = NULL,
  target_system = NULL,
  records_compared = 0,
  records_matched = 0,
  discrepancies_found = 0
)
}
\arguments{
\item{closeout_id}{Closeout ID}

\item{reconciliation_type}{Type}

\item{performed_by}{User performing}

\item{source_system}{Source system}

\item{target_system}{Target system}

\item{records_compared}{Number compared}

\item{records_matched}{Number matched}

\item{discrepancies_found}{Number of discrepancies}
}
\value{
List with success status
}
\description{
Record Data Reconciliation
}
\keyword{internal}
