% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_import.R
\name{redcap_extract_subjects}
\alias{redcap_extract_subjects}
\title{Extract REDCap subjects (record IDs).}
\usage{
redcap_extract_subjects(conn, pid)
}
\description{
Extract REDCap subjects (record IDs).
}
\keyword{internal}
