% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_import.R
\name{redcap_extract_subjects_api}
\alias{redcap_extract_subjects_api}
\title{Extract REDCap subject IDs via the REST API.}
\usage{
redcap_extract_subjects_api(api, metadata = NULL)
}
\description{
Extract REDCap subject IDs via the REST API.
}
\keyword{internal}
