% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_import.R
\name{redcap_extract_users}
\alias{redcap_extract_users}
\title{Extract REDCap users for a project.}
\usage{
redcap_extract_users(conn, pid)
}
\description{
Extract REDCap users for a project.
}
\keyword{internal}
