% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_import.R
\name{redcap_extract_users_api}
\alias{redcap_extract_users_api}
\title{Extract REDCap users via the REST API.}
\usage{
redcap_extract_users_api(api)
}
\description{
Extract REDCap users via the REST API.
}
\keyword{internal}
