% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_import.R
\name{redcap_translate_field_type}
\alias{redcap_translate_field_type}
\title{Translate a REDCap (field_type, validation_type) pair to a
ZZedc field_type.}
\usage{
redcap_translate_field_type(redcap_type, validation_type = NA)
}
\description{
Translate a REDCap (field_type, validation_type) pair to a
ZZedc field_type.
}
\keyword{internal}
