% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_choice_module.R
\name{render_setup_choice}
\alias{render_setup_choice}
\title{Show Setup Choice Page}
\usage{
render_setup_choice(configured = TRUE, id = "setup_choice")
}
\arguments{
\item{configured}{Boolean - is system already configured?}

\item{id}{Namespace ID}
}
\value{
UI or NULL
}
\description{
Conditionally render setup choice page if not configured
}
\keyword{internal}
