% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{resume_processing_activity}
\alias{resume_processing_activity}
\title{Resume Processing Activity}
\usage{
resume_processing_activity(activity_id, resumed_by, resume_reason)
}
\arguments{
\item{activity_id}{Activity ID}

\item{resumed_by}{User resuming the activity}

\item{resume_reason}{Reason for resuming (minimum 20 characters)}
}
\value{
List with success status
}
\description{
Resumes a previously stopped processing activity (if compelling grounds).
}
\keyword{internal}
