% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erasure.R
\name{safe_int_erasure}
\alias{safe_int_erasure}
\title{Safe Integer Scalar Conversion for Erasure}
\usage{
safe_int_erasure(x, default = NA_integer_)
}
\arguments{
\item{x}{Value to convert}

\item{default}{Default value if NULL}
}
\value{
Integer scalar or NA
}
\description{
Safe Integer Scalar Conversion for Erasure
}
\keyword{internal}
