% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectification.R
\name{safe_int_rect}
\alias{safe_int_rect}
\title{Safe Integer Scalar Conversion for Rectification}
\usage{
safe_int_rect(x, default = NA_integer_)
}
\arguments{
\item{x}{Value to convert}

\item{default}{Default value if NULL}
}
\value{
Integer scalar or NA
}
\description{
Safe Integer Scalar Conversion for Rectification
}
\keyword{internal}
