% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{setup_branching_logic}
\alias{setup_branching_logic}
\title{Setup branching logic handlers}
\usage{
setup_branching_logic(input, output, session, form_fields, form_id = "form")
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{form_fields}{List of field definitions with branching rules}

\item{form_id}{Namespace ID for the form}
}
\description{
Sets up reactive observers to handle field visibility changes.
Called once when the form is initialized.
}
\keyword{internal}
