% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_config.R
\name{sqlite_config_template}
\alias{sqlite_config_template}
\title{SQLite Configuration Template}
\usage{
sqlite_config_template()
}
\value{
Character string with YAML template
}
\description{
SQLite Configuration Template
}
\keyword{internal}
