% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{submit_change_request}
\alias{submit_change_request}
\title{Submit Change Request}
\usage{
submit_change_request(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with success status
}
\description{
Submit Change Request
}
\keyword{internal}
