% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_version.R
\name{unlock_crf_definition}
\alias{unlock_crf_definition}
\title{Unlock CRF Definition}
\usage{
unlock_crf_definition(crf_id, unlocked_by, unlock_reason)
}
\arguments{
\item{crf_id}{CRF ID}

\item{unlocked_by}{User unlocking the CRF}

\item{unlock_reason}{Reason for unlocking}
}
\value{
List with success status
}
\description{
Unlocks a previously locked CRF.
}
\keyword{internal}
