% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{unlock_database}
\alias{unlock_database}
\title{Unlock Database}
\usage{
unlock_database(study_id, unlocked_by, authorized_by, reason)
}
\arguments{
\item{study_id}{Study ID}

\item{unlocked_by}{User unlocking}

\item{authorized_by}{Authorizing user}

\item{reason}{Unlock reason}
}
\value{
List with success status
}
\description{
Unlock Database
}
\keyword{internal}
