% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccg.R
\name{update_ccg_form_status}
\alias{update_ccg_form_status}
\title{Update CCG Form Status}
\usage{
update_ccg_form_status(form_id, status, updated_by)
}
\arguments{
\item{form_id}{Form ID}

\item{status}{New status from get_ccg_form_statuses()}

\item{updated_by}{User updating the form}
}
\value{
List with success status
}
\description{
Updates the status of a CCG form.
}
\keyword{internal}
