% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{update_change_request_status}
\alias{update_change_request_status}
\title{Update Change Request Status}
\usage{
update_change_request_status(request_id, new_status)
}
\arguments{
\item{request_id}{Request ID}

\item{new_status}{New status}
}
\value{
List with success status
}
\description{
Update Change Request Status
}
\keyword{internal}
