% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{update_form_status}
\alias{update_form_status}
\title{Update Form Status}
\usage{
update_form_status(design_id, new_status, updated_by)
}
\arguments{
\item{design_id}{Design ID}

\item{new_status}{New status}

\item{updated_by}{User updating}
}
\value{
List with success status
}
\description{
Update Form Status
}
\keyword{internal}
