% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_templates.R
\name{use_template}
\alias{use_template}
\title{Use Template}
\usage{
use_template(template_id, used_by, study_name = NULL, crf_id = NULL)
}
\arguments{
\item{template_id}{Template ID}

\item{used_by}{User}

\item{study_name}{Optional study name}

\item{crf_id}{Optional CRF ID}
}
\value{
List with success status
}
\description{
Use Template
}
\keyword{internal}
