% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_module.R
\name{user_management_ui}
\alias{user_management_ui}
\title{User Management UI}
\usage{
user_management_ui(id)
}
\arguments{
\item{id}{The namespace id for the module}
}
\value{
A tagList containing the user management UI
}
\description{
User Management UI
}
\examples{
\dontrun{
# Embed the user and role management module.
# Most users will run the full application via
# `zzedc::launch_zzedc()`; the example below shows how to
# embed the module in a custom Shiny app instead.
if (interactive()) {
  ui <- shiny::fluidPage(
    user_management_ui("module_id")
  )
  server <- function(input, output, session) {
    user_management_server("module_id")
  }
  shiny::shinyApp(ui, server)
}
}
}
\keyword{internal}
