% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_validators.R
\name{validate_value_for_visit}
\alias{validate_value_for_visit}
\title{Check if value is valid for a visit}
\usage{
validate_value_for_visit(value, visit_code, allowed_visits)
}
\arguments{
\item{value}{Field value}

\item{visit_code}{Visit code}

\item{allowed_visits}{Vector of visits where this field should be completed}
}
\value{
TRUE if value valid for this visit
}
\description{
Some fields have different validation rules based on visit type.
Example: ADHD rating only required at specific visits
}
\keyword{internal}
