% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules_module.R
\name{validation_rules_server}
\alias{validation_rules_server}
\title{Validation Rules Server}
\usage{
validation_rules_server(id, user_info)
}
\arguments{
\item{id}{Module namespace ID}

\item{user_info}{Reactive containing user information}
}
\description{
Validation Rules Server
}
\keyword{internal}
