% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_gsheets_integration.R
\name{VALIDATION_SHEET_SCHEMA}
\alias{VALIDATION_SHEET_SCHEMA}
\title{Expected columns in the validation_rules Google Sheet}
\usage{
VALIDATION_SHEET_SCHEMA
}
\description{
Expected columns in the validation_rules Google Sheet
}
\keyword{internal}
