% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_history_module.R
\name{version_history_ui}
\alias{version_history_ui}
\title{Version History UI}
\usage{
version_history_ui(id)
}
\arguments{
\item{id}{Character: Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Creates the UI for the version history viewer module.
}
\keyword{internal}
