% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_validators.R
\name{visit_in_list}
\alias{visit_in_list}
\title{Check if visit is in a set of visits}
\usage{
visit_in_list(visit_code, visit_list)
}
\arguments{
\item{visit_code}{Current visit code}

\item{visit_list}{List of visit codes to check}
}
\value{
TRUE if visit_code is in visit_list
}
\description{
Checks if the current visit code matches any in a list.
Example: if visit in ('baseline', 'screening') then ...
}
\keyword{internal}
