% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.R
\name{activate_restriction_request}
\alias{activate_restriction_request}
\title{Activate Restriction Request}
\usage{
activate_restriction_request(request_id, activated_by)
}
\arguments{
\item{request_id}{Request ID}

\item{activated_by}{User activating the request}
}
\value{
List with success status
}
\description{
Activates a restriction request after all items have been applied.
}
\examples{
\dontrun{
activate_restriction_request(
  request_id = 1,
  activated_by = "dpo"
)
}
}
