% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{add_checklist_item}
\alias{add_checklist_item}
\title{Add Checklist Item}
\usage{
add_checklist_item(
  closeout_id,
  item_category,
  item_description,
  item_order = 1,
  is_required = TRUE
)
}
\arguments{
\item{closeout_id}{Closeout ID}

\item{item_category}{Category}

\item{item_description}{Description}

\item{item_order}{Display order}

\item{is_required}{Whether required}
}
\value{
List with success status
}
\description{
Add Checklist Item
}
