% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{add_condition_to_group}
\alias{add_condition_to_group}
\title{Add Condition to Group}
\usage{
add_condition_to_group(group_id, condition_id)
}
\arguments{
\item{group_id}{Group ID}

\item{condition_id}{Condition ID}
}
\value{
List with success status
}
\description{
Add Condition to Group
}
