% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{add_form_section}
\alias{add_form_section}
\title{Add Form Section}
\usage{
add_form_section(
  design_id,
  section_code,
  section_name,
  section_order = 1,
  is_repeating = FALSE,
  columns = 1
)
}
\arguments{
\item{design_id}{Design ID}

\item{section_code}{Section code}

\item{section_name}{Section name}

\item{section_order}{Order}

\item{is_repeating}{Whether repeating}

\item{columns}{Number of columns}
}
\value{
List with success status
}
\description{
Add Form Section
}
