% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\name{add_risk_assessment}
\alias{add_risk_assessment}
\title{Add Risk Assessment}
\usage{
add_risk_assessment(
  assessment_id,
  risk_category,
  risk_description,
  likelihood,
  impact,
  assessed_by,
  mitigation_measure = NULL,
  residual_risk = NULL
)
}
\arguments{
\item{assessment_id}{Assessment ID}

\item{risk_category}{Category}

\item{risk_description}{Description}

\item{likelihood}{Likelihood}

\item{impact}{Impact level}

\item{assessed_by}{User assessing}

\item{mitigation_measure}{Mitigation measure}

\item{residual_risk}{Residual risk}
}
\value{
List with success status
}
\description{
Add Risk Assessment
}
