% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{add_rule_to_set}
\alias{add_rule_to_set}
\title{Add Rule to Set}
\usage{
add_rule_to_set(set_id, rule_id, execution_order = NULL)
}
\arguments{
\item{set_id}{Set ID}

\item{rule_id}{Rule ID}

\item{execution_order}{Optional execution order}
}
\value{
List with success status
}
\description{
Add Rule to Set
}
