% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_templates.R
\name{add_template_field}
\alias{add_template_field}
\title{Add Field to Template}
\usage{
add_template_field(
  template_id,
  field_code,
  field_name,
  field_label,
  field_type,
  field_order,
  section_name = NULL,
  is_required = FALSE,
  field_lib_id = NULL,
  help_text = NULL
)
}
\arguments{
\item{template_id}{Template ID}

\item{field_code}{Field code}

\item{field_name}{Field name}

\item{field_label}{Field label}

\item{field_type}{Field type}

\item{field_order}{Display order}

\item{section_name}{Optional section}

\item{is_required}{Whether required}

\item{field_lib_id}{Optional link to library field}

\item{help_text}{Optional help text}
}
\value{
List with success status
}
\description{
Add Field to Template
}
