% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\docType{package}
\name{breach_notification}
\alias{zzedc}
\alias{breach_notification}
\title{Breach Notification Workflow (Feature #29)}
\description{
GDPR Articles 33-34 compliant breach detection, assessment,
notification, and documentation system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
