% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculate BMI}
\usage{
calculate_bmi(weight, height, decimal_places = 2)
}
\arguments{
\item{weight}{Weight in kg}

\item{height}{Height in cm}

\item{decimal_places}{Decimal places}
}
\value{
List with BMI value
}
\description{
Calculate BMI
}
