% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{calculate_score_sum}
\alias{calculate_score_sum}
\title{Calculate Score Sum}
\usage{
calculate_score_sum(values, na_handling = "ignore")
}
\arguments{
\item{values}{Vector of values to sum}

\item{na_handling}{How to handle NA (ignore, zero, error)}
}
\value{
List with sum
}
\description{
Calculate Score Sum
}
