% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\docType{package}
\name{calculated_fields}
\alias{calculated_fields}
\title{Calculated/Derived Fields (Feature #31)}
\description{
Auto-calculated fields with formula definitions for BMI,
age calculations, date differences, and custom formulas.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
