% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\docType{package}
\name{change_control}
\alias{change_control}
\title{Change Control System (Feature #27)}
\description{
Provides formal change control for CRFs, protocols, and system
modifications with impact assessment and approval workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
