% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\docType{package}
\name{conditional_logic}
\alias{conditional_logic}
\title{Conditional Logic & Dependencies (Feature #30)}
\description{
Show/hide logic, field dependencies, and conditional validation
for CRF forms and fields.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
