% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{create_validation_rule_set}
\alias{create_validation_rule_set}
\title{Create Validation Rule Set}
\usage{
create_validation_rule_set(
  set_code,
  set_name,
  created_by,
  set_description = NULL
)
}
\arguments{
\item{set_code}{Unique set code}

\item{set_name}{Set name}

\item{created_by}{User creating set}

\item{set_description}{Optional description}
}
\value{
List with success status
}
\description{
Create Validation Rule Set
}
