% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_wizard_utils.R
\name{create_wizard_directories}
\alias{create_wizard_directories}
\title{Create Directories for New ZZedc Instance}
\usage{
create_wizard_directories(base_path)
}
\arguments{
\item{base_path}{Base directory where subdirectories will be created}
}
\value{
List with success status
}
\description{
Creates the directory structure needed for a new ZZedc installation
}
