% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\docType{package}
\name{crf_designer}
\alias{crf_designer}
\title{WYSIWYG CRF Designer (Feature #32)}
\description{
Drag-and-drop CRF design interface with form sections,
field placement, and layout management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
