% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_review.R
\docType{package}
\name{crf_review}
\alias{crf_review}
\title{CRF Design Review Workflow System}
\description{
Implements structured review workflow for CRF designs with
multiple review stages, reviewer assignments, comments,
and approval tracking.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
