% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_templates.R
\docType{package}
\name{crf_templates}
\alias{crf_templates}
\title{CRF Template Library System}
\description{
Provides pre-built CRF templates for common clinical trial forms.
Templates can be customized and used as starting points for new CRFs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
