% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{delete_form_field}
\alias{delete_form_field}
\title{Delete Form Field}
\usage{
delete_form_field(field_id)
}
\arguments{
\item{field_id}{Field ID}
}
\value{
List with success status
}
\description{
Delete Form Field
}
