% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{evaluate_condition}
\alias{evaluate_condition}
\title{Evaluate Condition}
\usage{
evaluate_condition(source_value, operator, comparison_value)
}
\arguments{
\item{source_value}{Value to evaluate}

\item{operator}{Operator}

\item{comparison_value}{Value to compare}
}
\value{
Boolean result
}
\description{
Evaluate Condition
}
