% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_library.R
\docType{package}
\name{field_library}
\alias{field_library}
\title{Master Field Library System}
\description{
Centralized library of standard field definitions for reuse
across CRFs. Supports CDISC-compliant fields, custom fields,
and field templates with validation rules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
