% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{get_action_taken_categories}
\alias{get_action_taken_categories}
\title{Get Action Taken Categories}
\usage{
get_action_taken_categories()
}
\value{
Named list of action taken
}
\description{
Returns valid action taken categories.
}
