% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\name{get_breach_timeline}
\alias{get_breach_timeline}
\title{Get Breach Timeline}
\usage{
get_breach_timeline(incident_id)
}
\arguments{
\item{incident_id}{Incident ID}
}
\value{
List with timeline
}
\description{
Get Breach Timeline
}
