% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated_fields.R
\name{get_calculated_fields_statistics}
\alias{get_calculated_fields_statistics}
\title{Get Calculated Fields Statistics}
\usage{
get_calculated_fields_statistics()
}
\value{
List with statistics
}
\description{
Get Calculated Fields Statistics
}
