% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{get_causality_categories}
\alias{get_causality_categories}
\title{Get Causality Assessments}
\usage{
get_causality_categories()
}
\value{
Named list of causality categories
}
\description{
Returns causality assessment categories.
}
