% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccg.R
\name{get_ccg_forms}
\alias{get_ccg_forms}
\title{Get CCG Forms}
\usage{
get_ccg_forms(include_inactive = FALSE, form_category = NULL, status = NULL)
}
\arguments{
\item{include_inactive}{Include inactive forms}

\item{form_category}{Optional filter by category}

\item{status}{Optional filter by status}
}
\value{
List with success status and forms
}
\description{
Retrieves CCG forms.
}
