% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccg.R
\name{get_ccg_versions}
\alias{get_ccg_versions}
\title{Get CCG Versions}
\usage{
get_ccg_versions(form_id)
}
\arguments{
\item{form_id}{Form ID}
}
\value{
List with success status and versions
}
\description{
Retrieves version history for a CCG form.
}
