% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{get_change_request_types}
\alias{get_change_request_types}
\title{Get Change Request Types}
\usage{
get_change_request_types()
}
\value{
Named character vector
}
\description{
Get Change Request Types
}
