% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{get_closeout_status}
\alias{get_closeout_status}
\title{Get Closeout Status}
\usage{
get_closeout_status(closeout_id)
}
\arguments{
\item{closeout_id}{Closeout ID}
}
\value{
List with closeout details
}
\description{
Get Closeout Status
}
