% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{get_conditional_groups}
\alias{get_conditional_groups}
\title{Get Conditional Groups}
\usage{
get_conditional_groups(form_id = NULL)
}
\arguments{
\item{form_id}{Optional form filter}
}
\value{
List with groups
}
\description{
Get Conditional Groups
}
