% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_correction.R
\name{get_correction_reasons}
\alias{get_correction_reasons}
\title{Get Correction Reasons}
\usage{
get_correction_reasons()
}
\value{
Named list of correction reasons
}
\description{
Returns list of valid correction reasons with descriptions.
}
