% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_version.R
\name{get_crf_change_categories}
\alias{get_crf_change_categories}
\title{Get Change Categories}
\usage{
get_crf_change_categories()
}
\value{
Named character vector of change categories
}
\description{
Get Change Categories
}
