% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{get_crf_designer_statistics}
\alias{get_crf_designer_statistics}
\title{Get CRF Designer Statistics}
\usage{
get_crf_designer_statistics()
}
\value{
List with statistics
}
\description{
Get CRF Designer Statistics
}
