% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_templates.R
\name{get_crf_templates}
\alias{get_crf_templates}
\title{Get CRF Templates}
\usage{
get_crf_templates(template_category = NULL, include_inactive = FALSE)
}
\arguments{
\item{template_category}{Optional category filter}

\item{include_inactive}{Include inactive templates}
}
\value{
List with templates
}
\description{
Get CRF Templates
}
