% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_version.R
\name{get_crf_version_statistics}
\alias{get_crf_version_statistics}
\title{Get CRF Version Statistics}
\usage{
get_crf_version_statistics()
}
\value{
List with statistics
}
\description{
Returns statistics about CRF versions and changes.
}
