% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsar.R
\name{get_data_categories}
\alias{get_data_categories}
\title{Get Data Categories}
\usage{
get_data_categories()
}
\value{
Named list of data categories
}
\description{
Returns standard personal data categories per GDPR.
}
