% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{get_data_source_types}
\alias{get_data_source_types}
\title{Get Data Source Types}
\usage{
get_data_source_types()
}
\value{
Named character vector of data source types
}
\description{
Returns valid data source types for eligibility determination.
}
\examples{
sources <- get_data_source_types()
}
