% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{get_designer_form}
\alias{get_designer_form}
\title{Get Designer Form}
\usage{
get_designer_form(design_id)
}
\arguments{
\item{design_id}{Design ID}
}
\value{
List with form details
}
\description{
Get Designer Form
}
