% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erasure.R
\name{get_erasure_exceptions}
\alias{get_erasure_exceptions}
\title{Get Erasure Exceptions}
\usage{
get_erasure_exceptions()
}
\value{
Named list of exception grounds
}
\description{
Returns grounds for refusing erasure under GDPR Article 17(3).
}
