% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erasure.R
\name{get_erasure_grounds}
\alias{get_erasure_grounds}
\title{Get Erasure Grounds}
\usage{
get_erasure_grounds()
}
\value{
Named list of erasure grounds
}
\description{
Returns valid grounds for erasure under GDPR Article 17.
}
