% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_library.R
\name{get_field_library_statistics}
\alias{get_field_library_statistics}
\title{Get Field Library Statistics}
\usage{
get_field_library_statistics()
}
\value{
List with statistics
}
\description{
Returns statistics about the field library.
}
