% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{get_field_options}
\alias{get_field_options}
\title{Get Field Options}
\usage{
get_field_options(field_id)
}
\arguments{
\item{field_id}{Field ID}
}
\value{
List with options
}
\description{
Get Field Options
}
