% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_library.R
\name{get_field_usage}
\alias{get_field_usage}
\title{Get Field Usage}
\usage{
get_field_usage(field_lib_id)
}
\arguments{
\item{field_lib_id}{Field library ID}
}
\value{
List with success status and usage records
}
\description{
Retrieves usage history for a field.
}
