% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_logic.R
\name{get_group_conditions}
\alias{get_group_conditions}
\title{Get Group Conditions}
\usage{
get_group_conditions(group_id)
}
\arguments{
\item{group_id}{Group ID}
}
\value{
List with conditions
}
\description{
Get Group Conditions
}
