% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control.R
\name{get_impact_assessments}
\alias{get_impact_assessments}
\title{Get Impact Assessments}
\usage{
get_impact_assessments(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with assessments
}
\description{
Get Impact Assessments
}
