% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\name{get_lock_history}
\alias{get_lock_history}
\title{Get Lock History}
\usage{
get_lock_history(study_id)
}
\arguments{
\item{study_id}{Study ID}
}
\value{
List with lock history
}
\description{
Get Lock History
}
