% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{get_objectionable_legal_bases}
\alias{get_objectionable_legal_bases}
\title{Get Legal Bases for Objection}
\usage{
get_objectionable_legal_bases()
}
\value{
Named character vector of legal bases
}
\description{
Returns legal bases that can be objected to.
}
