% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{get_pending_objection_requests}
\alias{get_pending_objection_requests}
\title{Get Pending Objection Requests}
\usage{
get_pending_objection_requests()
}
\value{
List with success status and requests
}
\description{
Get Pending Objection Requests
}
