% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\name{get_pia_assessment}
\alias{get_pia_assessment}
\title{Get PIA Assessment}
\usage{
get_pia_assessment(assessment_id)
}
\arguments{
\item{assessment_id}{Assessment ID}
}
\value{
List with assessment details
}
\description{
Get PIA Assessment
}
