% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\name{get_pia_assessments}
\alias{get_pia_assessments}
\title{Get PIA Assessments List}
\usage{
get_pia_assessments(status = NULL)
}
\arguments{
\item{status}{Optional status filter}
}
\value{
List with assessments
}
\description{
Get PIA Assessments List
}
