% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{get_portability_datasets}
\alias{get_portability_datasets}
\title{Get Portability Datasets}
\usage{
get_portability_datasets(request_id, eligible_only = FALSE)
}
\arguments{
\item{request_id}{Request ID}

\item{eligible_only}{Only return eligible datasets (default FALSE)}
}
\value{
List with success status and datasets
}
\description{
Retrieves datasets for a portability request.
}
\examples{
\dontrun{
datasets <- get_portability_datasets(request_id = 1, eligible_only = TRUE)
}
}
