% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{get_portability_transfers}
\alias{get_portability_transfers}
\title{Get Portability Transfers}
\usage{
get_portability_transfers(request_id)
}
\arguments{
\item{request_id}{Request ID}
}
\value{
List with success status and transfers
}
\description{
Retrieves transfers for a portability request.
}
\examples{
\dontrun{
transfers <- get_portability_transfers(request_id = 1)
}
}
