% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectification.R
\name{get_rectification_statistics}
\alias{get_rectification_statistics}
\title{Get Rectification Statistics}
\usage{
get_rectification_statistics(db_path = NULL)
}
\arguments{
\item{db_path}{Character: Database path (optional)}
}
\value{
List with statistics
}
\description{
Returns rectification statistics.
}
