% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectification.R
\name{get_rectification_types}
\alias{get_rectification_types}
\title{Get Rectification Types}
\usage{
get_rectification_types()
}
\value{
Named list of rectification types
}
\description{
Returns valid rectification request types.
}
