% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.R
\name{get_restriction_scopes}
\alias{get_restriction_scopes}
\title{Get Restriction Scopes}
\usage{
get_restriction_scopes()
}
\value{
Named character vector of scope values
}
\description{
Returns valid scope values for restriction items.
}
\examples{
scopes <- get_restriction_scopes()
}
