% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{get_retention_action_history}
\alias{get_retention_action_history}
\title{Get Retention Action History}
\usage{
get_retention_action_history(record_id)
}
\arguments{
\item{record_id}{Record ID}
}
\value{
List with success status and history
}
\description{
Retrieves action history for a record.
}
