% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{get_retention_data_categories}
\alias{get_retention_data_categories}
\title{Get Data Categories}
\usage{
get_retention_data_categories()
}
\value{
Named character vector of data categories
}
\description{
Get Data Categories
}
