% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{get_retention_statistics}
\alias{get_retention_statistics}
\title{Get Retention Statistics}
\usage{
get_retention_statistics()
}
\value{
List with statistics
}
\description{
Returns comprehensive retention statistics.
}
