% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{get_retention_statuses}
\alias{get_retention_statuses}
\title{Get Retention Statuses}
\usage{
get_retention_statuses()
}
\value{
Named character vector of record statuses
}
\description{
Get Retention Statuses
}
