% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_review.R
\name{get_review_comment_severities}
\alias{get_review_comment_severities}
\title{Get Comment Severities}
\usage{
get_review_comment_severities()
}
\value{
Named character vector of severities
}
\description{
Get Comment Severities
}
