% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_review.R
\name{get_review_comment_types}
\alias{get_review_comment_types}
\title{Get Comment Types}
\usage{
get_review_comment_types()
}
\value{
Named character vector of comment types
}
\description{
Get Comment Types
}
