% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_review.R
\name{get_review_cycle_statuses}
\alias{get_review_cycle_statuses}
\title{Get Review Cycle Statuses}
\usage{
get_review_cycle_statuses()
}
\value{
Named character vector of cycle statuses
}
\description{
Get Review Cycle Statuses
}
