% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_review.R
\name{get_review_cycle_types}
\alias{get_review_cycle_types}
\title{Get Review Cycle Types}
\usage{
get_review_cycle_types()
}
\value{
Named character vector of cycle types
}
\description{
Get Review Cycle Types
}
