% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_review.R
\name{get_review_stage_types}
\alias{get_review_stage_types}
\title{Get Review Stage Types}
\usage{
get_review_stage_types()
}
\value{
Named character vector of stage types
}
\description{
Get Review Stage Types
}
