% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_review.R
\name{get_review_stages}
\alias{get_review_stages}
\title{Get Review Stages}
\usage{
get_review_stages(cycle_id)
}
\arguments{
\item{cycle_id}{Cycle ID}
}
\value{
List with success status and stages
}
\description{
Retrieves stages for a review cycle.
}
