% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{get_section_fields}
\alias{get_section_fields}
\title{Get Section Fields}
\usage{
get_section_fields(section_id)
}
\arguments{
\item{section_id}{Section ID}
}
\value{
List with fields
}
\description{
Get Section Fields
}
