% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electronic_signatures.R
\name{get_signature_statement}
\alias{get_signature_statement}
\title{Get Signature Statement}
\usage{
get_signature_statement(meaning)
}
\arguments{
\item{meaning}{Character: Signature meaning code}
}
\value{
Character: Signature statement
}
\description{
Returns the legal statement for a signature meaning.
}
