% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{get_subject_consents}
\alias{get_subject_consents}
\title{Get Subject Consents}
\usage{
get_subject_consents(subject_id, include_withdrawn = FALSE)
}
\arguments{
\item{subject_id}{Subject identifier}

\item{include_withdrawn}{Include withdrawn consents}
}
\value{
List with success status and consents
}
\description{
Gets all consents for a subject.
}
