% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectification.R
\name{get_third_party_recipients}
\alias{get_third_party_recipients}
\title{Get Third-Party Recipients}
\usage{
get_third_party_recipients(request_id, db_path = NULL)
}
\arguments{
\item{request_id}{Integer: Request ID}

\item{db_path}{Character: Database path (optional)}
}
\value{
Data frame with recipients
}
\description{
Retrieves third-party recipients for a request.
}
