% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{get_transfer_methods}
\alias{get_transfer_methods}
\title{Get Transfer Methods}
\usage{
get_transfer_methods()
}
\value{
Named character vector of transfer methods
}
\description{
Returns valid transfer methods for controller-to-controller transfers.
}
\examples{
methods <- get_transfer_methods()
}
