% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\name{get_validation_rule_sets}
\alias{get_validation_rule_sets}
\title{Get Rule Sets}
\usage{
get_validation_rule_sets(include_inactive = FALSE)
}
\arguments{
\item{include_inactive}{Include inactive sets}
}
\value{
List with rule sets
}
\description{
Get Rule Sets
}
