% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{get_withdrawal_scopes}
\alias{get_withdrawal_scopes}
\title{Get Withdrawal Scopes}
\usage{
get_withdrawal_scopes()
}
\value{
Named character vector of withdrawal scopes
}
\description{
Get Withdrawal Scopes
}
