% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.R
\name{lift_restriction_item}
\alias{lift_restriction_item}
\title{Lift Restriction Item}
\usage{
lift_restriction_item(item_id, lifted_by, lift_reason)
}
\arguments{
\item{item_id}{Item ID}

\item{lifted_by}{User lifting the restriction}

\item{lift_reason}{Reason for lifting (minimum 20 characters)}
}
\value{
List with success status
}
\description{
Lifts the restriction on a data item.
}
\examples{
\dontrun{
lift_restriction_item(
  item_id = 1,
  lifted_by = "dpo",
  lift_reason = "Accuracy dispute resolved in favor of data subject"
)
}
}
