% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_library.R
\name{load_cdisc_standard_fields}
\alias{load_cdisc_standard_fields}
\title{Load CDISC Standard Fields}
\usage{
load_cdisc_standard_fields(loaded_by)
}
\arguments{
\item{loaded_by}{User loading the fields}
}
\value{
List with success status and count
}
\description{
Loads standard CDISC SDTM fields into the library.
}
