% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_version.R
\name{lock_crf_definition}
\alias{lock_crf_definition}
\title{Lock CRF Definition}
\usage{
lock_crf_definition(crf_id, locked_by, lock_reason)
}
\arguments{
\item{crf_id}{CRF ID}

\item{locked_by}{User locking the CRF}

\item{lock_reason}{Reason for locking}
}
\value{
List with success status
}
\description{
Locks a CRF to prevent further modifications.
}
