% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{lock_form_design}
\alias{lock_form_design}
\title{Lock Form Design}
\usage{
lock_form_design(design_id, locked_by)
}
\arguments{
\item{design_id}{Design ID}

\item{locked_by}{User locking}
}
\value{
List with success status
}
\description{
Lock Form Design
}
