% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.R
\name{notify_restriction_third_party}
\alias{notify_restriction_third_party}
\title{Notify Restriction Third Party}
\usage{
notify_restriction_third_party(recipient_id, sent_by)
}
\arguments{
\item{recipient_id}{Recipient ID}

\item{sent_by}{User who sent the notification}
}
\value{
List with success status
}
\description{
Records notification sent to third party about restriction.
}
\examples{
\dontrun{
notify_restriction_third_party(
  recipient_id = 1,
  sent_by = "dpo"
)
}
}
