% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\docType{package}
\name{objection}
\alias{objection}
\title{Right to Object System}
\description{
GDPR Article 21 compliant objection to processing system.
Enables data subjects to object to processing based on legitimate
interest, direct marketing, or research purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
