% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\docType{package}
\name{portability}
\alias{portability}
\title{Right to Data Portability System}
\description{
GDPR Article 20 compliant data portability system.
Enables data subjects to receive their personal data in a structured,
commonly used, machine-readable format and transmit it to another controller.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
