% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_service.R
\name{prepare_sample_export}
\alias{prepare_sample_export}
\title{Prepare sample data for export}
\usage{
prepare_sample_export(options = NULL)
}
\arguments{
\item{options}{List of export options}
}
\value{
data.frame with sample data
}
\description{
Prepare sample data for export
}
