% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\docType{package}
\name{privacy_impact}
\alias{privacy_impact}
\title{Privacy Impact Assessment (PIA) Tool (Feature #28)}
\description{
GDPR Article 35 compliant privacy impact assessment system for
evaluating data processing activities and privacy risks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
