% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_linkage.R
\docType{package}
\name{protocol_linkage}
\alias{protocol_linkage}
\title{Protocol-CRF Linkage System (Feature #25)}
\description{
Links CRF fields and forms to protocol requirements, ensuring
traceability between protocol objectives and data collection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
