% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portability.R
\name{record_export_download}
\alias{record_export_download}
\title{Record Export Download}
\usage{
record_export_download(export_id, downloaded_by)
}
\arguments{
\item{export_id}{Export ID}

\item{downloaded_by}{User downloading}
}
\value{
List with success status
}
\description{
Records when an export file is downloaded.
}
\examples{
\dontrun{
record_export_download(export_id = 1, downloaded_by = "subject")
}
}
