% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach_notification.R
\name{report_breach_incident}
\alias{report_breach_incident}
\title{Report Breach Incident}
\usage{
report_breach_incident(
  incident_title,
  breach_type,
  severity,
  description,
  detected_by,
  breach_start = NULL,
  data_categories_affected = NULL,
  subjects_affected = 0,
  records_affected = 0
)
}
\arguments{
\item{incident_title}{Title}

\item{breach_type}{Type}

\item{severity}{Severity}

\item{description}{Description}

\item{detected_by}{User detecting}

\item{breach_start}{When breach started}

\item{data_categories_affected}{Categories affected}

\item{subjects_affected}{Number of subjects}

\item{records_affected}{Number of records}
}
\value{
List with success status
}
\description{
Report Breach Incident
}
