% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.R
\docType{package}
\name{restriction}
\alias{restriction}
\title{Right to Restrict Processing System}
\description{
GDPR Article 18 compliant restriction of processing system.
Allows data subjects to request that their data not be processed
while still being retained. Integrates with legal holds and
provides third-party notification per Article 19.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
