% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objection.R
\name{stop_processing_activity}
\alias{stop_processing_activity}
\title{Stop Processing Activity}
\usage{
stop_processing_activity(activity_id, stopped_by)
}
\arguments{
\item{activity_id}{Activity ID}

\item{stopped_by}{User stopping the activity}
}
\value{
List with success status
}
\description{
Stops a processing activity due to objection.
}
