% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_closeout.R
\docType{package}
\name{study_closeout}
\alias{study_closeout}
\title{Study Reconciliation & Closeout System (Feature #26)}
\description{
Provides study closeout procedures, data reconciliation,
database lock management, and closeout checklists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
