% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_impact.R
\name{submit_pia_for_review}
\alias{submit_pia_for_review}
\title{Submit PIA for Review}
\usage{
submit_pia_for_review(assessment_id)
}
\arguments{
\item{assessment_id}{Assessment ID}
}
\value{
List with success status
}
\description{
Submit PIA for Review
}
