% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_version.R
\name{update_crf_version_status}
\alias{update_crf_version_status}
\title{Update Version Status}
\usage{
update_crf_version_status(version_id, status, updated_by)
}
\arguments{
\item{version_id}{Version ID}

\item{status}{New status from get_crf_version_statuses()}

\item{updated_by}{User updating status}
}
\value{
List with success status
}
\description{
Updates the status of a CRF version.
}
