% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_designer.R
\name{update_field_position}
\alias{update_field_position}
\title{Update Field Position}
\usage{
update_field_position(
  field_id,
  field_row = NULL,
  field_column = NULL,
  section_id = NULL
)
}
\arguments{
\item{field_id}{Field ID}

\item{field_row}{New row}

\item{field_column}{New column}

\item{section_id}{New section}
}
\value{
List with success status
}
\description{
Update Field Position
}
