% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_management_module.R
\name{user_management_server}
\alias{user_management_server}
\title{User Management Server}
\usage{
user_management_server(id, db_pool = NULL)
}
\arguments{
\item{id}{The namespace id for the module}

\item{db_pool}{Reactive expression returning database connection pool}
}
\value{
List of reactive expressions and functions
}
\description{
User Management Server
}
\examples{
\dontrun{
# Embed the user and role management module.
# Most users will run the full application via
# `zzedc::launch_zzedc()`; the example below shows how to
# embed the module in a custom Shiny app instead.
if (interactive()) {
  ui <- shiny::fluidPage(
    user_management_ui("module_id")
  )
  server <- function(input, output, session) {
    user_management_server("module_id")
  }
  shiny::shinyApp(ui, server)
}
}
}
\keyword{internal}
