% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_rules.R
\docType{package}
\name{validation_rules}
\alias{validation_rules}
\title{Advanced Validation Rules System}
\description{
Provides a domain-specific language (DSL) for defining and executing
data validation rules for clinical trial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzedc}
  \item Report bugs at \url{https://github.com/rgt47/zzedc/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu}

}
