% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_gsheets_integration.R
\docType{data}
\name{VALIDATION_SHEET_SCHEMA}
\alias{VALIDATION_SHEET_SCHEMA}
\title{Expected columns in the validation_rules Google Sheet}
\format{
An object of class \code{list} of length 3.
}
\usage{
VALIDATION_SHEET_SCHEMA
}
\description{
Expected columns in the validation_rules Google Sheet
}
\keyword{internal}
